update lsr to take uint32 and float

+++ lsr/lsr.h
@@ -19,6 +19,6 @@
 
 /* returns linear root of least squares linear regression of series of `len'
  * number of (`time', `level') ordered pairs */
-float lsr_root(float *time, float *level, int len);
+float lsr_root(uint32_t *time, float *level, int len);
 
 #endif /* LSR_H */
